# api/openapi.yaml
openapi: "3.1.0"
info:
  title: EncodeSwarmr API
  version: "1.0.0"
  description: >
    REST API for the EncodeSwarmr controller. Manages encoding jobs,
    agents, sources, templates, variables, webhooks, users, and audit logs.
    All versioned endpoints are prefixed with /api/v1.
servers:
  - url: /api/v1
    description: Versioned API base path
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      summary: List script templates
      tags: [templates]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by template type (e.g. avs, vpy, bat)
          schema:
            type: string
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createTemplate
      summary: Create a script template
      description: Requires admin role.
      tags: [templates]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Template"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /templates/{id}:
    get:
      operationId: getTemplate
      summary: Get a template by ID
      tags: [templates]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateTemplate
      summary: Update a template
      description: Requires admin role.
      tags: [templates]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateRequest"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Template"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteTemplate
      summary: Delete a template
      description: Requires admin role.
      tags: [templates]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Template deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Variables ─────────────────────────────────────────────────────────────
  /variables:
    get:
      operationId: listVariables
      summary: List global variables
      tags: [variables]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
      responses:
        "200":
          description: Variable list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Variable"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /variables/{name}:
    get:
      operationId: getVariable
      summary: Get a variable by name
      tags: [variables]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Variable
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Variable"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: upsertVariable
      summary: Create or update a variable
      description: Requires admin role.
      tags: [variables]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVariableRequest"
      responses:
        "200":
          description: Variable upserted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Variable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /variables/{id}:
    delete:
      operationId: deleteVariable
      summary: Delete a variable by ID
      description: Requires admin role.
      tags: [variables]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Variable deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Path Mappings ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: HTTP-only session cookie set by POST /auth/login or OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: API key sent as a Bearer token in the Authorization header.

  # ─── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Insufficient role
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Validation error (unprocessable entity)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    Envelope:
      type: object
      properties:
        data:
          description: The primary response payload.
        meta:
          $ref: "#/components/schemas/Meta"
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
          description: Correlation identifier echoed from X-Request-ID.
...
    Problem:
      description: RFC 9457 Problem Details
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string

    # Domain schemas
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          description: Template file type (e.g. avs, vpy, bat).
        extension:
          type: string
          example: avs
        content:
          type: string
          description: Go template body with {{.Variable}} placeholders.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Variable:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateTemplateRequest:
      type: object
      required: [name, type, extension, content]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          description: Template type (avs, vpy, bat).
        extension:
          type: string
          example: avs
        content:
          type: string
          description: Go template body.
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
    UpsertVariableRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
        description:
          type: string
        category:
          type: string
...